+++ changelog.md
@@ -1,3 +1,10 @@
+# Version 1.73.0 (2024-11-22)
+
+## What's Changed
+- Bump hahomematic to 2024.11.9
+  - Make sysvars eventable
+- Add option to receive sysvar changes over mqtt
+
 # Version 1.72.0 (2024-11-21)
 
 ## What's Changed